Fix encoder dtype promotion: _topk_sparse .float() cast z to float32
The mask = (torch.abs(z) >= threshold).float() promoted the latent
code from float16 to float32, causing decoder forward pass to fail.
Changed to .to(dtype=z.dtype) to preserve the input dtype.

+++ src/project_adam/encoder.py
@@ -68,7 +68,7 @@ class SensoryEncoder(nn.Module):
         k = max(1, int(z.shape[-1] * _SPARSITY_KEEP_FRAC))
         vals, _ = torch.topk(torch.abs(z), k, dim=-1)
         threshold = vals[..., -1].unsqueeze(-1)
-        mask = (torch.abs(z) >= threshold).float()
+        mask = (torch.abs(z) >= threshold).to(dtype=z.dtype)
         return z * mask
 
     def forward(self, x):
